+++ kubejs/server_scripts/create_factory_logistics/tags.js
@@ -1,4 +1,5 @@
 // priority: 0
+"use strict";
 
 function registerCreateFactoryLogisticsItemTags(event) {
 	event.add('c:hidden_from_recipe_viewers', 'create_factory_logistics:incomplete_fluid_mechanism')