+++ kubejs/server_scripts/createdeco/recipes.js
@@ -1,3 +1,5 @@
+"use strict";
+
 const registerCreatedecoRecipes = (event) => {
 
 	//#region Item Replacements