+++ kubejs/server_scripts/tfc/recipes.materials.js
@@ -1,4 +1,5 @@
 ﻿// priority: 0
+"use strict";
 
 function registerTFCMaterialsRecipes(event) {