+++ kubejs/server_scripts/tfcbetterbf/data.js
@@ -1,4 +1,5 @@
 // priority: 0
+"use strict";
 
 const registerTFCDataForTFCBetterBF = (event) => {
     event.itemHeat('tfcbetterbf:insulation', 8.571, null, null)