+++ kubejs/server_scripts/tfg/recipes.food.js
@@ -1,4 +1,6 @@
 ﻿// priority: 0
+"use strict";
+
 /**
  * @param {Internal.RecipesEventJS} event 
  */