+++ kubejs/startup_scripts/computer_craft/constants.js
@@ -1,4 +1,5 @@
 // priority: 0
+"use strict";
 
 global.COMPUTER_CRAFT_DISCS = [
     '{Color:15790320}',