+++ kubejs/startup_scripts/extended_ae2/constants.js
@@ -1,3 +1,5 @@
+"use strict";
+
 global.EXTENDED_AE2_DISABLED_ITEMS = [
 
     'expatternprovider:infinity_cell',