+++ kubejs/startup_scripts/modern_markings/constants.js
@@ -1,4 +1,5 @@
 // priority: 0
+"use strict";
 
 global.MODERN_MARKINGS_DISABLED_ITEMS = [
     'ags_modernmarkings:chalk'